requirements.txt
genanki==0.13.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
Pillow>=10.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
